+++ lib/main.dart
@@ -1,130 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-class WelcomeScreen extends StatelessWidget {
-  const WelcomeScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [
-              Color.fromARGB(255, 93, 63, 104),
-              Color.fromARGB(255, 62, 39, 80),
-              Color.fromARGB(255, 49, 33, 59),
-            ],
-          ),
-        ),
-        child: SafeArea(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 32.0),
-            child: Column(
-              children: [
-                const Spacer(flex: 2),
-                
-                // Título de boas-vindas
-                SizedBox(
-                  width: double.infinity,
-                  child: Text(
-                    'Welcome to\nListfy!',
-                    textAlign: TextAlign.left,
-                    style: TextStyle(
-                      fontSize: 48,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.white,
-                      height: 1.2,
-                      fontFamily: GoogleFonts.chivo().fontFamily,
-                    ),
-                  ),
-                ),
-                
-                const Spacer(flex: 3),
-            
-                SizedBox(
-                  width: double.infinity,
-                  height: 56,
-                  child: ElevatedButton(
-                    onPressed: () {
-                    },
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.black,
-                      foregroundColor: Colors.white,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(28),
-                      ),
-                      elevation: 0,
-                    ),
-                    child: Text(
-                      'Sign up',
-                      style: TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.w600,
-                        fontFamily: GoogleFonts.chivo().fontFamily,
-                      ),
-                    ),
-                  ),
-                ),
-                
-                const SizedBox(height: 16),
-                
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      'Already have an account? ',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 16,
-                        fontFamily: GoogleFonts.chivo().fontFamily,
-                      ),
-                    ),
-                    GestureDetector(
-                      onTap: () {
-                      },
-                      child:                       Text(
-                        'Login',
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 16,
-                          fontWeight: FontWeight.w600,
-                          decoration: TextDecoration.underline,
-                          decorationColor: Colors.white,
-                          fontFamily: GoogleFonts.chivo().fontFamily,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-                
-                const Spacer(flex: 1),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Listfy',
-      theme: ThemeData(
-        primarySwatch: Colors.purple,
-      ),
-      home: const WelcomeScreen(),
-      debugShowCheckedModeBanner: false,
-    );
-  }
-}
+import 'app.dart';
 
 void main() {
   runApp(const MyApp());
